中国地图显示了

+++ src/VicoldUtility4Net5/VicoldGis/VMap/Projections/IProjection.cs
@@ -7,6 +7,6 @@ namespace VicoldGis.VMap.Projections
 {
     internal interface IProjection
     {
-        public Point Project(float x, float y);
+        public Point Project(double x, double y);
     }
 }
